+++ odoo/local-src/stock_delivery_note/tests/test_delivery_note.py
@@ -111,6 +111,10 @@ class TestStockDeliveryNote(SavepointCase):
         # do new transfer must be done in setUp as
         # setUpClass doesn't set odoo.tools.config['test_enable']
         self.picking.do_new_transfer()
+        # This is a hack because the reserved_quant_id are not filled up
+        # And they are needed by the get_lot in stock.move
+        for sm in self.picking.move_lines:
+            sm.linked_move_operation_ids[0].reserved_quant_id = sm.quant_ids[0]
 
     def test_delivery_note_filename(self):
         """Check the correct generation of the filename"""